Pagination view class and pagination for leaderboards cog
